MySQL tracing without USDT moved to dbslower
